refactor(spark): extract shared base classes for HoodiePartitionValues
Move the InternalRow delegation methods shared by all Spark versions from
Spark3HoodiePartitionValues and the per-version Spark 4.x copies into
BaseHoodiePartitionValues in hudi-spark-common, and the getVariant
delegation shared by all Spark 4.x versions into an abstract
Spark4HoodiePartitionValues in hudi-spark4-common. The Spark 4.0/4.1/4.2
classes now only carry copy() plus, on 4.1/4.2, the geography/geometry
getters introduced by Spark 4.1.
